long

+++ dataset.py
@@ -36,7 +36,7 @@ class MultimodalDataset(Dataset):
             path_imgs = os.path.join(data_dir, 'images', cls + '.t7')
             path_txts = os.path.join(data_dir, 'text_c10', cls + '.t7')
             cls_imgs = torch.Tensor(torch.load(path_imgs).float())
-            cls_txts = torch.LongTensor(torch.load(path_txts))
+            cls_txts = torch.LongTensor(torch.load(path_txts).long())
             self.data[cls] = (cls_imgs, cls_txts)
 
             self.num_instances += cls_imgs.size(0)
